+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [2.1.1](https://github.com/pchuri/confluence-cli/compare/v2.1.0...v2.1.1) (2026-04-28)
+
+
+### Bug Fixes
+
+* strip callout markers and narrow detection to paragraph start ([#132](https://github.com/pchuri/confluence-cli/issues/132)) ([8bec8b3](https://github.com/pchuri/confluence-cli/commit/8bec8b36c98a9ef1929760aef313471f05bf10d4))
+
+# [2.1.0](https://github.com/pchuri/confluence-cli/compare/v2.0.2...v2.1.0) (2026-04-28)
+
+
+### Features
+
+* **converter:** add EXPAND marker for collapsible expand macro ([#130](https://github.com/pchuri/confluence-cli/issues/130)) ([efe0677](https://github.com/pchuri/confluence-cli/commit/efe067736e0abec034e08b4052f0375e4e066a94))
+
 ## [2.0.2](https://github.com/pchuri/confluence-cli/compare/v2.0.1...v2.0.2) (2026-04-27)